Extract frontend executable functionality to a separate contextual class.
Fixes #7771
