+++ flow-maven-plugin/src/main/java/com/vaadin/flow/plugin/maven/BuildFrontendMojo.java
@@ -41,6 +41,7 @@ import com.vaadin.flow.component.dependency.JsModule;
 import com.vaadin.flow.component.dependency.NpmPackage;
 import com.vaadin.flow.server.Constants;
 import com.vaadin.flow.server.ExecutionFailedException;
+import com.vaadin.flow.server.frontend.FrontendTools;
 import com.vaadin.flow.server.frontend.FrontendUtils;
 import com.vaadin.flow.server.frontend.NodeTasks;
 import com.vaadin.flow.theme.Theme;
@@ -181,12 +182,13 @@ public class BuildFrontendMojo extends FlowModeAbstractMojo {
         }
 
         String nodePath;
+        FrontendTools tools = new FrontendTools(npmFolder.getAbsolutePath(),
+                ()-> FrontendUtils.getVaadinHomeDirectory().getAbsolutePath());
         if (requireHomeNodeExec) {
-            nodePath = FrontendUtils
-                    .ensureNodeExecutableInHome(npmFolder.getAbsolutePath());
+            nodePath = tools
+                    .forceAlternativeNodeExecutable();
         } else {
-            nodePath = FrontendUtils
-                    .getNodeExecutable(npmFolder.getAbsolutePath());
+            nodePath = tools.getNodeExecutable();
         }
 
         List<String> command = Arrays.asList(nodePath,
