+++ flow-maven-plugin/src/main/java/com/vaadin/flow/plugin/maven/PrepareFrontendMojo.java
@@ -37,6 +37,7 @@ import org.sonatype.plexus.build.incremental.BuildContext;
 
 import com.vaadin.flow.server.Constants;
 import com.vaadin.flow.server.ExecutionFailedException;
+import com.vaadin.flow.server.frontend.FrontendTools;
 import com.vaadin.flow.server.frontend.FrontendUtils;
 import com.vaadin.flow.server.frontend.NodeTasks;
 
@@ -95,8 +96,10 @@ public class PrepareFrontendMojo extends FlowModeAbstractMojo {
         propagateBuildInfo();
 
         try {
-            FrontendUtils
-                    .validateNodeAndNpmVersion(npmFolder.getAbsolutePath());
+            FrontendTools tools = new FrontendTools(npmFolder.getAbsolutePath(),
+                    () -> FrontendUtils.getVaadinHomeDirectory()
+                            .getAbsolutePath());
+            tools.validateNodeAndNpmVersion();
         } catch (IllegalStateException exception) {
             throw new MojoExecutionException(exception.getMessage(), exception);
         }
