+++ flow-server/src/test/java/com/vaadin/flow/server/frontend/TaskRunPnpmInstallTest.java
@@ -45,9 +45,11 @@ public class TaskRunPnpmInstallTest extends TaskRunNpmInstallTest {
     public void setUp() throws IOException {
         super.setUp();
         // ensure there is a valid pnpm installed in the system
-        FrontendUtils.ensurePnpm(getNodeUpdater().npmFolder.getAbsolutePath());
+        new FrontendTools(getNodeUpdater().npmFolder.getAbsolutePath(), null)
+                .ensurePnpm();
         // create an empty package.json so as pnpm can be run without error
-        FileUtils.write(new File(getNodeUpdater().npmFolder, PACKAGE_JSON), "{}", StandardCharsets.UTF_8);
+        FileUtils.write(new File(getNodeUpdater().npmFolder, PACKAGE_JSON),
+                "{}", StandardCharsets.UTF_8);
     }
 
     @Test
